+++ src/StackExchange.Redis/Enums/RedisCommand.cs
@@ -23,6 +23,7 @@ internal enum RedisCommand
     CLUSTER,
     CONFIG,
     COPY,
+    COMMAND,
 
     DBSIZE,
     DEBUG,
@@ -86,6 +87,7 @@ internal enum RedisCommand
 
     LASTSAVE,
     LATENCY,
+    LCS,
     LINDEX,
     LINSERT,
     LLEN,
@@ -360,6 +362,7 @@ internal static class RedisCommandExtensions
             case RedisCommand.BITPOS:
             case RedisCommand.CLIENT:
             case RedisCommand.CLUSTER:
+            case RedisCommand.COMMAND:
             case RedisCommand.CONFIG:
             case RedisCommand.DBSIZE:
             case RedisCommand.DEBUG:
@@ -393,6 +396,7 @@ internal static class RedisCommandExtensions
             case RedisCommand.KEYS:
             case RedisCommand.LASTSAVE:
             case RedisCommand.LATENCY:
+            case RedisCommand.LCS:
             case RedisCommand.LINDEX:
             case RedisCommand.LLEN:
             case RedisCommand.LPOS: