+++ src/StackExchange.Redis/RawResult.cs
@@ -266,6 +266,9 @@ namespace StackExchange.Redis
         internal string?[]? GetItemsAsStrings() => this.ToArray<string?>((in RawResult x) => (string?)x.AsRedisValue());
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal string[]? GetItemsAsStringsNotNullable() => this.ToArray<string>((in RawResult x) => (string)x.AsRedisValue()!);
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal bool[]? GetItemsAsBooleans() => this.ToArray<bool>((in RawResult x) => (bool)x.AsRedisValue());
 
         internal GeoPosition? GetItemsAsGeoPosition()